feat(quadrangulate): substrate fold repair — Winslow untangler (auto), fold census, QGP relinearization measured off
The Bi-MDF tracer consumes the relaxed reduced map, whose hard seam
constraints invert triangles at cones (fold diagnosis 2026-08; measured
here at quarter density: nefertiti@4000 pure 482 folds at reduced-preint,
89% cone-adjacent, vs 28 at poisson). Three additions, all inert on the
default path:

- CYBER_FOLD_DIAG fold census (permanent, gated): per-phase folded-face
  counts with cone-distance attribution (d0/d1/d2/3+) at poisson, arap,
  reduced-preint, reduced-repaired and final, via a FoldRepairContext
  (per-face cut triples, reference areas/shapes, cone BFS distances)
  built only when the census or the Bi-MDF substrate is active.

- QGP §7.1 dynamic re-linearization (CYBER_QC_FOLD_RELIN, default OFF by
  measurement): convexified positive-area rows as soft penalties on the
  reduced solve, re-linearized about each new pose. Three formulations
  measured on nefertiti@4000 pure: the QGP-faithful monotone active set
  cascades (folds 482 -> 7211 — the equality pulls recovered faces back
  to the margin); normalized rows explode along near-floppy directions
  of the reduced operator (6400-cell displacements at any lambda; rows
  are restricted to continuous frees for the same reason); the stable
  raw-row/relative-weight variant only reaches 482 -> 382 and its
  margin-slivers break the tracer (collinear separatrix overlap).
  Rewinding a >2pi wedge at a negative cone is intrinsically nonlinear.

- Regularized-Winslow substrate untangling (Garanzha et al. 2021,
  CYBER_QC_FOLD_UNTANGLE, default auto): minimizes the chi-regularized
  Winslow energy over the reduced free basis w (z = Tuv w keeps every
  iterate exactly seamless; the reduction transports all wedge copies
  consistently, subsuming the per-cone fixed-boundary patch design),
  L-BFGS with Armijo and the paper's eps-schedule, double precision.
  Substrate-only: the untangled map feeds the T-mesh tracer; w is
  restored so the rounding path is untouched. Folds 482 -> 45
  (nefertiti pure, 0.46s), 198 -> 9 (armadillo pure), 41 -> 0 (bunny).
  auto engages above 1% folded faces only: always-on cost the healthy
  bunny flagship ears 19 -> 20 / sing 94 -> 100 for zero benefit.

Supporting changes: bimdf::Charts u/v are now double — a float substrate
splits exactly-coincident twin separatrix levels by one ulp (measured
dc=5.96e-8), landing in the tracer's collinear-overlap hazard band
[1e-9,1e-7); float sources widen exactly, so the raw path is unchanged.
CYBER_QC_BIMDF_HEALTH is now a full measurement override (also skips the
crease-fraction refusal — nefertiti's 293 crease arcs otherwise block
engagement regardless of health). CYBER_QC_BIMDF_STEERTOL (default 0)
optionally drops steering rows whose substrate arc length disagrees with
the realized relaxed length — on the healthy bunny even one skipped arc
moved ears 19 -> 22, so it stays an A/B lever. A gated debug print
identifies colliding separatrix levels under CYBER_QC_BIMDF_DEBUG.

Chain re-measure on nefertiti@4000 pure: contained regions 371 -> 314,
sideMismatch 0.731 -> 0.627, twinMerges 491 -> 247, but degraded nodes
110 -> 125 — classification does not fully clear. Gate (pure sing <=
200): NOT met — best 333 (geometry-damaged) / 358 (geometry-sane,
multires+guided+untangle) vs greedy 419; the measured residue is the
cone population (1297 field cones; CYBER_QC_CROSSFIELD_MULTIRES removes
only 29), redirecting future work to field-level dipole annihilation.

Gates: ctest 14/14; bench check green off/report/guided; bunny default
guided ears 19 / sing 94 exact; box_sharp@1000 pure exact injection
(maxFrac 0.0000); CAD guided spot-check vs main 16/16 metric-SAME;
organics default guided vs main 5/5 metric-SAME; off-vs-report 18/18
metric-SAME. Byte-hash identity gates are retired on this machine: the
main-branch binary itself is run-to-run non-byte-deterministic (metrics
identical to 4 decimals).

+++ src/quadrangulate/src/bimdf_quantize.hpp
@@ -52,8 +52,12 @@ struct SeamEdge {
 // The cut-open relaxed-seamless parameterization, in compact form.
 struct Charts {
     std::size_t nCut = 0;
-    const float* u = nullptr;  // relaxed-seamless UV per cut vertex (z = Tuv w)
-    const float* v = nullptr;
+    // Relaxed-seamless UV per cut vertex (z = Tuv w). Double so that exactly
+    // coincident separatrix levels (twin launches) stay coincident: a float
+    // substrate breaks them apart by one ulp, which the tracer's collinear-
+    // overlap hazard band rejects.
+    const double* u = nullptr;
+    const double* v = nullptr;
     std::vector<std::array<std::size_t, 3>> faces;  // per-face cut-vertex ids
     std::vector<int> coneIndex;                     // per cut vertex, 0 = regular
     std::vector<std::uint32_t> vertexOfCut;         // mesh vertex id per cut vertex
